[migrations] - Modificando las migraciones de  perfil y telefono_perfil para permitir ejecutar las migraciones posteriores

+++ database/migrations/2025_05_20_043745_create_perfil_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('perfil', function (Blueprint $table) {
-            $table->integer('idperfil')->primary();
+            $table->unsignedBigInteger('idperfil')->primary();
             $table->string('nombre');
             $table->string('apellido');
             $table->dateTime('fecha_nacimiento');
